% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accrual.R
\name{accrual}
\alias{accrual}
\alias{accrual_months}
\title{Accrual plot}
\usage{
accrual(
  actual,
  expected,
  target = sum(expected),
  col = par("col"),
  pch = 18L,
  type = "b",
  time0 = 0,
  lag = 0,
  start = "Month 0",
  end = "Last month",
  full = "Full accrual expected",
  pos.full = 2L,
  pct = c(1, 0.75, 0.25, 0),
  fill = c("yellow", "orange", "red"),
  pct.legend = "Percent of expected",
  args.pct = list(),
  legend = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  ylab2 = NULL,
  axes = TRUE,
  ...
)

accrual_months(dates)
}
\arguments{
\item{actual, expected}{integer vectors of actual and expected for each time
unit; note these are expected to be sequential with no gaps, so units with
0 accrued should be included}

\item{target}{target accrual}

\item{time0}{accrual at time 0; default is 0}

\item{lag}{time units to shift the accrual line}

\item{start, end, full}{labels for first, last, and full accrual times; if
\code{NA} or \code{NULL}, these will not be added}

\item{pos.full}{for the \code{full} label, \code{pos} value pased to
\code{\link[text]}}

\item{pct, fill}{colored (\code{fill}) bands for accrual ranges based on
expected accrual; use \code{NA} or \code{NULL} to suppress}

\item{pct.legend}{label for \code{pct} legend; use \code{NA} or \code{NULL}
to suppress}

\item{args.pct}{a \emph{named} list of arguments passed to \code{\link{legend}}
controlling the \code{pct.legend}}

\item{legend}{optional \emph{named} vector to annotate plot; if \code{TRUE},
one will be made; use \code{NA} or \code{NULL} to suppress; see examples}

\item{xlim, ylim}{the x- and y-axis limits}

\item{xlab, ylab, ylab2}{the x-, y-, and secondary y-axis labels}

\item{axes}{logical; if \code{TRUE}, plot axes and box are drawn}

\item{...}{additional arguments passed to \code{plot}}

\item{dates}{a vector of \code{\link{Dates}} or a class to be coerced}

\item{, }{col,pch,type arguments controlling the accrual line, passed to
\code{\link{points}}}
}
\value{
A list with the \code{x} and \code{y} coordinates of the accrual line.
Additionally, the coordinates for polygon(s) for \code{pct}, if given.
}
\description{
Plot (unitless but usually monthly) accrual numbers. \code{accrual_months}
is a helper function to convert a vector of dates into the proper format
for \code{actual}.
}
\examples{
## basic usage
set.seed(1)
ac <- rpois(12, 2)
ex <- rep(c(1, 3, 5), times = c(1, 3, 7))
accrual(ac, ex)
accrual(ac, ex, start = 'Jan 2000', legend = TRUE, fill = adjustcolor(1:3, 0.5))
accrual(ac, ex, start = 'Jan 2000', args.pct = list(x = 'topleft'),
        pct = c(1.25, 1, 0.75, 0.5, 0.25), fill = topo.colors(4))

## minimal plot
accrual(ac, ex, pct = NA, start = NA, end = NA, full = NA, xlab = 'Time',
        axes = FALSE, panel.first = {box(); axis(1); axis(2)}, ylim = c(0, 50))

## adding multiple accrual lines
accrual(ac, ex)
accrual(ac * 0.75, col = 'blue')
accrual(ac[1:4], col = 'blue', lag = 6)
points(c(0, seq_along(ac)), c(0, cumsum(ac)), cex = 2)

## adding legends
st <- 'May 2025'
en <- 'Apr 2026'
lg <- c(
  'Start Date' = st,
  'Expected' = sum(ex), 'Exp. end date' = en, 'Exp. per month' = '2-4',
  'Total to date' = sprintf('\%s (\%.0f\%\%)', sum(ac), sum(ac) / sum(ex) * 100),
  'N per month' = round(mean(ac), 1)
)
op <- par(mar = c(5, 5, 2, 5), las = 1L)
accrual(ac, ex, legend = lg, start = st, end = en, fill = c(3, 7, 2),
        pos.full = 1L)
par(op)

## using a vector of dates
x <- Sys.Date() + unlist(sapply(seq_along(ac), function(ii)
  30 * ii + rpois(ac[ii], 5)))
## these are equivalent ways to call accrual with date vectors
a1 <- accrual(x, ex)
a2 <- accrual(accrual_months(x), ex)
identical(a1, a2)

## the plot data is returned
plot(
  a1$x, a1$y, ylim = c(0, max(unlist(a1))),
  panel.first = matlines(a1$pct$x, a1$pct[, -1])
)

}
