% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_text.R
\name{b2p_text}
\alias{b2p_text}
\title{Two-proportion design text}
\usage{
b2p_text(p1, p2, n1 = NULL, n2 = n1, power = NULL, alpha = 0.025,
  r = 0.5, arms = c("Experimental", "Control"), type = c("fisher",
  "UMPU", "approx.cor", "approx.unc"), cont.cor = TRUE)
}
\arguments{
\item{p1, p2}{the success rates for both groups (\code{p1 > p2})}

\item{n1, n2}{sample size for both groups}

\item{power, alpha}{the power and one-sided significance level (type I error)}

\item{r}{proportion assigned to group 1}

\item{arms}{labels for the two groups}

\item{type}{the type of test used to calculate power, see
\code{\link[desmon]{b2p}}}

\item{cont.cor}{logical; if \code{TRUE} (default), the sample size will be
calculated for the continuity corrected statistic}

\item{conf}{confidence level for single- and two-stage confidence intervals}

\item{outcome}{text string describing the outcome; if a string such as
\code{"long description (abbr)"} is given, then entire string is used in
the first instance, and only the text in parens is used subsequently}
}
\value{
A character string describing the two-stage design with
\code{attr(., "design")} containing the design.
}
\description{
Output two-proportion design text. The power or sample size (per group)
will be calculated depending on the arguments given (i.e., either
\code{n1} or \code{power} should be \code{NULL}).
}
\examples{
## basic usage (compare to a simon design)
p1 <- 0.5
p2 <- 0.3

b2p_text(p1, p2, power = 0.8)
b2p_text(p1, p2, n1 = 103)
b2p_text(p1, p2, power = 0.8, r = 2/3)

\dontrun{
## use cat and/or strwrap for improved formatting and writing
cat(strwrap(b2p_text(p1, p2, power = 0.8), width = 80), sep = '\\n')

cat(b2p_text(p1, p2, power = 0.8), file = '~/simon_design.txt')
}

}
\seealso{
\code{\link[desmon]{b2p}}; \code{\link[desmon]{b2n}}

Other designs: \code{\link{bin1samp_text}},
  \code{\link{mtd_text}}, \code{\link{simon_text}},
  \code{\link{twostg_text}}
}
\concept{designs}
