% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desmon-extra.R
\name{bin1samp_power}
\alias{bin1samp_power}
\title{Power for one-sample exact binomial designs}
\usage{
bin1samp_power(p0, pa, n, r)
}
\arguments{
\item{p0, pa}{probability of success under the null and alternative
hypotheses, respectively}

\item{n}{sample size}

\item{r}{critical value}
}
\value{
A vector with the following elements:

\item{\code{type1}}{the overall type-I error}
\item{\code{type2}}{the overall type-II error}
}
\description{
Determines the power and significance level for a one-sided, one-sample
exact binomial test.
}
\examples{
p0 <- 0.1
pa <- 0.3
des <- desmon2:::bin1samp(p0, pa)

bin1samp_power(p0, pa, des['n'], des['r'] + 1L)

## compare
des[c('size', 'type2')]

}
\seealso{
\code{\link{twostg_power}}
}
