% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_text.R
\name{bin1samp_text}
\alias{bin1samp_text}
\title{One-sample design text}
\usage{
bin1samp_text(p0, pa, ..., conf = 0.95, outcome = "OUTCOME (OUT)")
}
\arguments{
\item{p0, pa}{the null and alternative hypotheses}

\item{...}{additional arguments passed to \code{\link[desmon]{bin1samp}}}

\item{conf}{confidence level for confidence intervals}

\item{outcome}{text string describing the outcome; if a string such as
\code{"long description (abbr)"} is given, then entire string is used in
the first instance, and only the text in parens is used subsequently}
}
\value{
A character string describing the design with \code{attr(, "design")}
containing the design.
}
\description{
Output one-sided, one-sample, single-stage exact binomial design text.
}
\examples{
## basic usage
bin1samp_text(0.3, 0.5)

\dontrun{
## use cat and/or strwrap for improved formatting and writing
cat(strwrap(bin1samp_text(.3, .5), width = 80), sep = '\\n')

cat(bin1samp_text(.3, .5), file = '~/bin1samp_design.txt')

cat(bin1samp_text(.3, .5, outcome = 'overall response (OR) rate'),
    file = '~/bin1samp_design2.txt')
}

}
\seealso{
\code{\link[desmon]{bin1samp}}; \code{\link[desmon]{binci}}

Other designs: \code{\link{b2p_text}},
  \code{\link{mtd_text}}, \code{\link{simon_text}},
  \code{\link{twostg_text}}
}
\concept{designs}
