% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{blt_table}
\alias{blt_table}
\title{BLT table}
\usage{
blt_table(r, n, prior = c(1, 1)/2, alpha = 0.025, cval = 1/3, digits = 2L)
}
\arguments{
\item{r, n}{number of DLTs observed and total size}

\item{prior}{a vector of length two giving the distribution assumptions
for a beta prior (default is \code{beta(0.5, 0.5)})}

\item{alpha}{type-I error probability}

\item{cval}{for specifed proportions of DLTs, the probability that the rate
will be greater than \code{cval} is computed}

\item{digits}{number of digits to print}
}
\value{
A vector of five elements with the total size, number of DLTs, \code{alpha}
critical value, \code{1 - alpha} critical value, and calculated probability
that the DLT rate is greater than \code{cval}.
}
\description{
Bayes (dose)-limiting toxicities table for designing phase I dose-
escalation studies.
}
\examples{
x <- seq(0.005, 0.995, length.out = 500)
plot(x, dbeta(x, 2, 3), type = 'l')

blt_table(3, 6, alpha = .0125)
t(Vectorize(blt_table)(0:8, 8))

}
\seealso{
\code{\link{dlt_table}}; \code{\link{pr_table}}
}
