% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctgov.R
\name{ctae}
\alias{ctae}
\alias{ctae_table}
\alias{ctae_report}
\title{Clinical trials adverse events}
\usage{
ctae(
  data,
  id = "casenum",
  toxdesc = "toxdesc",
  toxcat = "toxcat",
  comment = NULL
)

ctae_table(x, sae, atrisk)

ctae_report(
  data,
  id = "casenum",
  toxdesc = "toxdesc",
  toxcat = "toxcat",
  comment = NULL,
  sae,
  atrisk,
  strata = NULL
)
}
\arguments{
\item{data}{a data frame with adverse events}

\item{id, toxdesc, toxcat, comment}{column names with patient IDs, toxicity
descriptions/terms, toxicity categories/organ systems, and comments/additional
details (e.g., clarifications for other/specify)}

\item{sae}{logical; if \code{TRUE}, result is formatted for reporting SAEs;
otherwise, non-SAEs are assumed}

\item{atrisk}{number of patients at-risk (for each stratum)}

\item{strata}{optional column name defining groups of patients; note that
for \code{ctae_report}, each unique level of \code{strata} should have
an \code{atrisk} value}
}
\description{
Summarize and format adverse events for ct.gov/FDA reporting.
}
\examples{
dd <- data.frame(
  casenum = rep(1:2, c(3, 5)),
  toxdesc = c('aa', 'aa', 'b', 'aa', 'aaa', 'b', 'bb', 'c'),
  toxcat = c('A', 'A', 'B', 'A', 'A', 'B', 'B', 'C')
)
ae <- ctae(dd)
ae
c1 <- ctae_table(ae, FALSE, 20)
c1
## alternatively
c2 <- ctae_report(dd, sae = FALSE, atrisk = 20)
c2
identical(c1, c2)

## stratify table to report by cohort
dd$group <- factor(dd$casenum, 1:2, c('Arm A', 'Arm B'))
ctae_report(dd, sae = FALSE, atrisk = c(10, 10), strata = 'group')

}
