% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{dlt_table}
\alias{dlt_table}
\title{DLT table}
\usage{
dlt_table(prob, digits = getOption("digits"))
}
\arguments{
\item{prob}{a vector of probabilities for event}

\item{digits}{number of digits past the decimal point to keep}
}
\description{
Create a standard 3+3 dose-limiting toxicity table with probabilities of
dose-escalation for true but unknown probabilities of experiencing
toxicity.
}
\examples{
## probabilities of experiencing event
prob <- c(1, 5, 1:5 * 10) / 100

dlt_table(prob)
t(dlt_table(prob, 3))

}
\seealso{
\code{\link{blt_table}}; \code{\link{pr_table}}
}
