% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_text.R
\name{mtd_text}
\alias{mtd_text}
\title{MTD design text}
\usage{
mtd_text(prob = 1:5/10, ndose = 3L, expansion = 10L, digits = 2L)
}
\arguments{
\item{prob}{a vector of toxicity event rates}

\item{ndose}{number of dose levels}

\item{expansion}{size of the expansion cohort}

\item{digits}{number of digits past the decimal point to keep}
}
\value{
A character string describing the design with \code{attr(, "table")}
containing the DLT table in markdown format.
}
\description{
Output standard 3+3 maximum tolerated dose design text.
}
\examples{
## basic usage
mtd_text()
cat(mtd_text(1:10 / 20, 5, 15))

\dontrun{
cat(mtd_text(), file = '~/mtd_design.txt')
}

}
\seealso{
\code{\link{dlt_table}}; \code{\link{sim3p3}}

Other designs: \code{\link{b2p_text}},
  \code{\link{bin1samp_text}}, \code{\link{simon_text}},
  \code{\link{twostg_text}}
}
\concept{designs}
