% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{pr_table}
\alias{pr_table}
\title{Probability table}
\usage{
pr_table(prob, n, crit, greater = FALSE, digits = 2L)
}
\arguments{
\item{prob}{a vector of probabilities for event}

\item{n, crit}{sample size and critical number of events, respectively; note
that the probabilities will always be calculated as weakly less or greater
than \code{crit}, i.e., \code{Pr(>= crit)} or \code{Pr(<= crit)} depending
if \code{greater} is \code{TRUE} or \code{FALSE}}

\item{greater}{\code{logical}; the direction of \code{crit}: if
\code{FALSE}, the probabilities are calculated for \code{crit}
\emph{or fewer} events; if \code{TRUE}, the probabilities are calculated
for \code{crit} \emph{or more} events}

\item{digits}{number of digits past the decimal point to keep}
}
\description{
A single-stage function for calculating the probabilities of \code{crit}
or fewer (greater if \code{greater = TRUE}) events for true but unknown
\code{prob}abilities of the event occurring.
}
\examples{
prob <- c(1, 5, 1:5 * 10) / 100

## probabilities of crit or fewer events given true but unknown prob
pr_table(prob, n = 15, crit = 3, greater = FALSE)

## probabilities of crit or more events given true but unknown prob
pr_table(prob, n = 15, crit = 3, greater = TRUE)

## compare
1 - as.numeric(pr_table(prob, n = 15, crit = 2, greater = FALSE)[2L, ])

}
\seealso{
\code{\link{blt_table}}; \code{\link{dlt_table}}
}
