% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_text.R
\name{print.stat_text}
\alias{print.stat_text}
\title{Stat text printing method}
\usage{
\method{print}{stat_text}(x, width = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{"stat_text"}}

\item{width}{a positive integer giving the target column for wrapping
lines in the output; see \code{\link{strwrap}}}

\item{...}{ignored}
}
\description{
Stat text printing method
}
