% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim3p3}
\alias{sim3p3}
\title{Simulate 3+3 design}
\usage{
sim3p3(probs, expansion = 10)
}
\arguments{
\item{probs}{vector of true rates of DLTs}

\item{expansion}{(optional) integer giving size of expansion cohort}
}
\value{
A list with the following components:

\item{\code{$data}}{dose by patient matrix: 0 - no DLT, 1 - DLT}
\item{\code{$mtd}}{numeric dose level selected}
\item{\code{$level}}{integer mtd level (ie, row of \code{data} selected)}
\item{\code{$expansion}}{observed rate of DLT in expansion cohort}
\item{\code{$cohort}}{observed rate of DLT in all patients}
\item{\code{$enrolled}}{total number of patients enrolled on study}
}
\description{
Simulate a standard 3+3 dose-escalation study.
}
\examples{
## standard 3+3 with 10 patient expansion cohort
pr <- 1:5 / 10
sim3p3(pr)

x <- replicate(1000L, sim3p3(pr)$level)
barplot(table(x), xlab = 'RP2D')

}
\seealso{
\code{\link{simcrm}}
}
