% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_text.R
\name{simon_text}
\alias{simon_text}
\title{Simon design text}
\usage{
simon_text(p0, pa, ..., conf = 0.95, which = 1L, outcome = "OUTCOME (OUT)")
}
\arguments{
\item{p0, pa}{the null and alternative hypotheses}

\item{...}{additional arguments passed to \code{\link[desmon]{simon}}}

\item{conf}{confidence level for single- and two-stage confidence intervals}

\item{which}{optional; an integer selecting the design to use if multiple
are found}

\item{outcome}{text string describing the outcome; if a string such as
\code{"long desc (abbr)"} is given, then entire string is used in the first
instance, and only the text in parens is used subsequently}
}
\value{
A character string describing the Simon design with
\code{attr(, "design")} containing the designs.
}
\description{
Output Simon two-stage design text.
}
\examples{
## basic usage
p0 <- 0.3
pa <- 0.5
simon_text(p0, pa)

## select a different design
simon_text(p0, pa, which = 2)

\dontrun{
## use cat and/or strwrap for improved formatting and writing
cat(strwrap(simon_text(p0, pa), width = 80), sep = '\n')

cat(simon_text(p0, pa), file = '~/simon_design.txt')

cat(simon_text(p0, pa, outcome = 'overall response (OR) rate'),
    file = '~/simon_design2.txt')
}

}
\seealso{
\code{\link[desmon]{simon}}; \code{\link[desmon]{twostg}};
\code{\link[desmon]{twocon}}; \code{\link[desmon]{binci}}

Other designs: 
\code{\link{b2p_text}()},
\code{\link{bin1samp_text}()},
\code{\link{mtd_text}()},
\code{\link{twostg_text}()}
}
\concept{designs}
