% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desmon-extra.R
\name{twostg_power}
\alias{twostg_power}
\title{Power for two-stage designs}
\usage{
twostg_power(p0, pa, n1, n2, r1, r2)
}
\arguments{
\item{p0, pa}{probability of success under the null and alternative
hypotheses, respectively}

\item{n1, n2}{sample size of first and second stage}

\item{r1, r2}{maximum number of responses in first stage and overall where
treatment is declared ineffective}
}
\value{
A vector with the following elements:

\item{\code{Pr.stop1.H0}}{probability of stopping after the first stage
if \code{p0} is true}
\item{\code{Pr.stop1.H1}}{probability of stopping after the first stage
if \code{pa} is true}
\item{\code{type1}}{the overall type-I error}
\item{\code{type2}}{the overall type-II error}
\item{\code{E.tot.n.H0}}{expected total sample size if \code{p0} is true}
\item{\code{E.tot.n.H1}}{expected total sample size if \code{pa} is true}
}
\description{
Determines the operating characteristics of single-arm, two-stage designs.
}
\examples{
p0 <- 0.1
pa <- 0.3
des <- desmon2:::simon(p0, pa)$designs[1L, ]

twostg_power(p0, pa, des[['n1']], des[['n2']], des[['r1']], des[['r2']])

## compare
des

}
\seealso{
\code{\link{bin1samp_power}}
}
