% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_text.R
\name{twostg_text}
\alias{twostg_text}
\title{Two-stage design text}
\usage{
twostg_text(p0, pa, n1, n2, r1, r2, conf = 0.95,
  outcome = "OUTCOME (OUT)")
}
\arguments{
\item{p0, pa}{the null and alternative hypotheses}

\item{n1, n2}{sample size of first and second stage}

\item{r1, r2}{maximum number of responses in first stage and overall where
treatment is declared ineffective}

\item{conf}{confidence level for single- and two-stage confidence intervals}

\item{outcome}{text string describing the outcome; if a string such as
\code{"long description (abbr)"} is given, then entire string is used in
the first instance, and only the text in parens is used subsequently}
}
\value{
A character string describing the two-stage design with
\code{attr(., "design")} containing the design.
}
\description{
Output two-stage design text.
}
\examples{
## basic usage (compare to a simon design)
p0 <- 0.3
pa <- 0.5
s <- as.list(desmon2:::simon(p0, pa)$designs[1L, ])

twostg_text(p0, pa, s$n1, s$n2, s$r1, s$r2)

\dontrun{
## use cat and/or strwrap for improved formatting and writing
cat(strwrap(twostg_text(p0, pa, s$n1, s$n2, s$r1, s$r2), width = 80),
    sep = '\\n')

cat(twostg_text(p0, pa, s$n1, s$n2, s$r1, s$r2),
    file = '~/simon_design.txt')

cat(twostg_text(p0, pa, s$n1, s$n2, s$r1, s$r2,
                outcome = 'overall response (OR) rate'),
    file = '~/simon_design2.txt')
}

}
\seealso{
\code{\link{twostg_power}}; \code{\link[desmon]{simon}};
\code{\link[desmon]{twostg}}; \code{\link[desmon]{twocon}};
\code{\link[desmon]{binci}}

Other designs: \code{\link{b2p_text}},
  \code{\link{bin1samp_text}}, \code{\link{mtd_text}},
  \code{\link{simon_text}}
}
\concept{designs}
