% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desmon-extra.R
\name{bin1pow}
\alias{bin1pow}
\title{Power for one-sample exact binomial tests}
\usage{
bin1pow(n, p0, pa, r)
}
\arguments{
\item{n}{sample size}

\item{p0, pa}{probability of success under the null and alternative
hypotheses, respectively}

\item{r}{critical value}
}
\description{
Determines the power and significance level for a one-sided, one-sample
exact binomial test.
}
\examples{
p0 <- 0.1
pa <- 0.3

des <- desmon2:::bin1samp(p0, pa)
des[c('size', 'type2')]

bin1pow(des['n'], p0, pa, des['r'] + 1L)

}
