% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctgov.R
\name{ctae}
\alias{ctae}
\alias{ctae_table}
\title{Clinical trials adverse events}
\usage{
ctae(
  data,
  id = "casenum",
  toxdesc = "toxdesc",
  toxcat = "toxcat",
  comment = NULL
)

ctae_table(x, sae, atrisk)
}
\arguments{
\item{data}{a data frame with adverse events}

\item{id, toxdesc, toxcat, comment}{column names with patient IDs, toxicity
descriptions/terms, toxicity categories/organ systems, and comments/additional
details (e.g., clarifications for other/specify)}

\item{sae}{logical; if \code{TRUE}, result is formatted for reporting SAEs}

\item{atrisk}{number of patients at-risk}
}
\description{
Summarize and format adverse events for ct.gov/FDA reporting.
}
\examples{
dd <- data.frame(
  casenum = rep(1:2, c(3, 5)),
  toxdesc = c('aa', 'aa', 'b', 'aa', 'aaa', 'b', 'bb', 'c'),
  toxcat = c('A', 'A', 'B', 'A', 'A', 'B', 'B', 'C')
)
ae <- ctae(dd)
ae
ctae_table(ae, FALSE, 20)

}
