% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim3p3}
\alias{sim3p3}
\title{Simulate 3+3 design}
\usage{
sim3p3(probs, expansion = 10, escalation = TRUE, n.max = Inf)
}
\arguments{
\item{probs}{vector of true rates of DLTs}

\item{expansion}{(optional) integer giving size of expansion cohort}

\item{escalation}{logical; if \code{TRUE}, dose-escalation rules are used;
otherwise, de-escalation rules are used with decreasing \code{probs}}

\item{n.max}{maximum number of patients available for dose-finding and
expansion cohorts combined}
}
\value{
A list with the following components:

\item{\code{$data}}{dose by patient matrix: 0 - no DLT, 1 - DLT}
\item{\code{$mtd}}{numeric dose level selected}
\item{\code{$level}}{integer mtd level (ie, row of \code{data} selected)}
\item{\code{$expansion}}{observed rate of DLT in expansion cohort}
\item{\code{$cohort}}{observed rate of DLT in all patients}
\item{\code{$enrolled}}{total number of patients enrolled on study}
}
\description{
Simulate a standard 3+3 dose (de-)escalation study.
}
\examples{
## standard 3+3 with 10 patient expansion cohort
pr <- 1:5 / 10
sim3p3(pr)


## run 1000 3+3 studies
sim1 <- replicate(1000L, sim3p3(pr), simplify = FALSE)

op <- par(mfrow = c(1, 2), las = 1L)
out <- sapply(sim1, function(x) c(x$level, x$enrolled))
barplot(
  prop.table(table(out[1, ])),
  main = 'Dose selected',
  xlab = 'RP2D',
  ylab = 'Proportion selected'
)
boxplot(
  split(out[2, ], out[1, ]),
  ylim = c(0, length(pr) * 6 + 10),
  main = 'N used to select dose x',
  xlab = 'Dose selected\\n(0 - no safe dose)',
  ylab = 'Total size'
)
par(op)


## 3+3 de-escalation, starting at highest dose
## run 1000 3+3 studies
pr2 <- rev(pr)
sim2 <- replicate(1000L, sim3p3(pr2, escalation = FALSE), simplify = FALSE)

op <- par(mfrow = c(1, 2), las = 1L)
out <- sapply(sim2, function(x) c(x$level, x$enrolled))
barplot(
  prop.table(table(out[1, ])),
  main = 'Dose selected',
  xlab = 'RP2D',
  ylab = 'Proportion selected'
)
boxplot(
  split(out[2, ], out[1, ]),
  ylim = c(0, length(pr) * 6 + 10),
  main = 'N used to select dose x',
  xlab = 'Dose selected\\n(0 - no safe dose)',
  ylab = 'Total size'
)
par(op)

}
\seealso{
\code{\link{simcrm}}
}
