% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simcrm}
\alias{simcrm}
\title{Simulate CRM design}
\usage{
simcrm(probs, expansion = 0L, target = 1/6, nptmax = length(probs) * 6L, ...)
}
\arguments{
\item{probs}{vector of true rates of DLTs}

\item{expansion}{(optional) integer giving size of expansion cohort}

\item{target}{target AE rate}

\item{nptmax}{max number of patients used on study (not including the
expansion cohort)}

\item{...}{additional arguments passed to \code{\link[UBCRM]{simCrm}}}
}
\value{
A list with following components:

\item{\code{$data}}{\code{UBCRM::simCrm()$data}}
\item{\code{$mtd}}{numeric dose level selected}
\item{\code{$level}}{integer mtd level (ie, row of \code{data} selected)}
\item{\code{$expansion}}{observed rate of DLT in expansion cohort}
\item{\code{$cohort}}{observed rate of DLT in all patients}
\item{\code{$enrolled}}{total number of patients enrolled on study}
}
\description{
Simulate a \link[=simCrm]{CRM} dose-escalation study.
}
\examples{
pr <- 1:5 / 10
simcrm(pr)


## run 100 crm studies
sim1 <- replicate(100L, simcrm(pr), simplify = FALSE)

op <- par(mfrow = c(1, 2), las = 1L)
out <- sapply(sim1, function(x) c(x$level, x$enrolled))
barplot(
  prop.table(table(out[1, ])),
  main = 'Dose selected',
  xlab = 'RP2D',
  ylab = 'Proportion selected'
)
boxplot(
  split(out[2, ], out[1, ]),
  ylim = c(0, length(pr) * 6 + 10),
  main = 'N used to select dose x',
  xlab = 'Dose selected\n(0 - no safe dose)',
  ylab = 'Total size'
)
par(op)

}
\seealso{
\code{\link{sim3p3}}
}
