% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desmon-extra.R
\name{twostg_power}
\alias{twostg_power}
\alias{twostg_sim}
\title{Power for two-stage designs}
\usage{
twostg_power(p0, pa, n1, n2, r1, r2)

twostg_sim(
  p0,
  pa,
  n1,
  n2,
  r1 = seq.int(n1),
  r2 = seq.int(n1 + n2),
  plot = TRUE
)
}
\arguments{
\item{p0, pa}{probability of success under the null and alternative
hypotheses, respectively}

\item{n1, n2}{sample size of first and second stage}

\item{r1, r2}{maximum number of responses in first stage and overall where
treatment is declared ineffective}

\item{plot}{logical; if \code{TRUE}, the type-I and type-II errors are
plotted for each combination of (valid) \code{r1} and \code{r2} values}
}
\value{
\code{twostg_power} returns vector with the following elements:

\item{\code{Pr.stop1.H0}}{probability of stopping after the first stage
if \code{p0} is true}
\item{\code{Pr.stop1.H1}}{probability of stopping after the first stage
if \code{pa} is true}
\item{\code{type1}}{the overall type-I error}
\item{\code{type2}}{the overall type-II error}
\item{\code{E.tot.n.H0}}{expected total sample size if \code{p0} is true}
\item{\code{E.tot.n.H1}}{expected total sample size if \code{pa} is true}

\code{twostg_sim} returns a data frame with columns for each of the above
plus the following:

\item{\code{r1}}{critical values for the first stage}
\item{\code{r2}}{critical values for the second stage}
}
\description{
Determines the operating characteristics of single-arm, two-stage designs.
}
\examples{
p0 <- 0.1
pa <- 0.3
des <- desmon2:::simon(p0, pa)$designs[1L, ]
twostg_power(p0, pa, des[['n1']], des[['n2']], des[['r1']], des[['r2']])

## compare
des


## simulate over critical values
twostg_sim(p0, pa, des[['n1']], des[['n2']])

\dontrun{
res <- twostg_sim(p0, pa, des[['n1']], des[['n2']])
with(res, {
  iplotr::iscatter(
    type1, type2, group = type1 < 0.1 & type2 < 0.2,
    labels = list(
      r1 = r1, r2 = r2,
      alpha = round(type1, 3), power = round(1 - type2, 3)
    )
  )
})
}

}
\seealso{
\code{\link{bin1samp_power}}; \code{\link{bin1samp_sim}}
}
